requirements.txt
tensorflow
numpy
pandas
matplotlib
dvc
notebook
seaborn
python-box==6.0.2
pyYAML # to read yaml files with Python
tqdm
ensure==1.0.2
joblib
types-pyYAML
scipy
Flask
Flask-Cors

# -e .